requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.41.3
botocore==1.41.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-ckeditor==6.3.2
django-components==0.143.2
django-debug-toolbar==6.1.0
django-extensions==4.1
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-ninja==1.5.0
django-storages==1.14.6
django-user-agents==0.4.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
frozenlist==1.4.1
django-rosetta==0.10.2
idna==3.6
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.0.5
mysqlclient==2.2.7
pillow==10.2.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rpds-py==0.29.0
s3transfer==0.15.0
six==1.17.0
sqlparse==0.4.4
twilio==9.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==0.18.0
uritemplate==4.2.0
urllib3==2.2.1
user-agents==2.2.0
whitenoise==6.6.0
yarl==1.9.4
